+++ wp-content/plugins/events-calendar-pro/src/Tribe/Recurrence/Queue_Processor.php
@@ -3,7 +3,7 @@ class Tribe__Events__Pro__Recurrence__Queue_Processor {
 	const SCHEDULED_TASK = 'tribe_events_pro_process_recurring_events';
 
 	/**
-	 * Number of event instances to be processed in a single batch.
+	 *Number of event instances to be processed in a single batch.
 	 *
 	 * @var int
 	 */
@@ -82,6 +82,8 @@ class Tribe__Events__Pro__Recurrence__Queue_Processor {
 	 *
 	 * @param int $event_id
 	 * @param int $batch_size
+	 *
+	 * @return bool
 	 */
 	public function process_batch( $event_id, $batch_size = null ) {
 		/**
@@ -94,7 +96,8 @@ class Tribe__Events__Pro__Recurrence__Queue_Processor {
 		$this->batch_size = ( null === $batch_size ) ? $default_batch_size : (int) $batch_size;
 
 		$this->current_event_id = (int) $event_id;
-		$this->do_processing();
+
+		return $this->do_processing();
 	}
 
 	/**
@@ -197,6 +200,10 @@ class Tribe__Events__Pro__Recurrence__Queue_Processor {
 				break;
 			}
 
+			if ( ! $this->current_queue->have_ownership_of_job() ) {
+				return;
+			}
+
 			Tribe__Events__Pro__Recurrence__Meta::delete_unexcluded_event( $instance_id, $start_date );
 
 			unset( $instances_to_delete[ $instance_id ] );
@@ -216,6 +223,10 @@ class Tribe__Events__Pro__Recurrence__Queue_Processor {
 				break;
 			}
 
+			if ( ! $this->current_queue->have_ownership_of_job() ) {
+				return;
+			}
+
 			$instance = new Tribe__Events__Pro__Recurrence__Instance( $this->current_event_id, $date_duration, $instance_id );
 			$instance->save();
 
@@ -228,9 +239,19 @@ class Tribe__Events__Pro__Recurrence__Queue_Processor {
 
 	protected function do_creations() {
 		$exclusions = $this->current_queue->instances_to_exclude();
-		$instances_to_create = $this->current_queue->instances_to_create();
 
-		foreach ( $instances_to_create as $key => $date_duration ) {
+		$instances_to_create = array_values( $this->current_queue->instances_to_create() );
+
+		try {
+			$sequence = new Tribe__Events__Pro__Recurrence__Sequence( $instances_to_create, $this->current_event_id );
+		} catch ( Exception $e ) {
+			$exception = new Tribe__Exception( $e );
+			$exception->handle();
+
+			return;
+		}
+
+		foreach ( $sequence->get_sorted_sequence() as $key => $date_duration ) {
 			// Don't process more than the current batch size allows
 			if ( $this->batch_complete() ) {
 				break;
@@ -239,15 +260,23 @@ class Tribe__Events__Pro__Recurrence__Queue_Processor {
 			// Some instances may deliberately have been removed - let's remove
 			// them from the list of events to create and move on
 			if ( in_array( $date_duration, $exclusions ) ) {
-				unset( $instances_to_create[ $key ] );
+				unset( $instances_to_create[ $date_duration['original_index'] ] );
 				$this->processed++;
 				continue;
 			}
 
-			$instance = new Tribe__Events__Pro__Recurrence__Instance( $this->current_event_id, $date_duration );
-			$instance->save();
+			if ( ! $this->current_queue->have_ownership_of_job() ) {
+				return;
+			}
 
-			unset( $instances_to_create[ $key ] );
+			$sequence_number = isset( $date_duration['sequence'] ) ? $date_duration['sequence'] : 1;
+			$instance        = new Tribe__Events__Pro__Recurrence__Instance( $this->current_event_id, $date_duration, 0, $sequence_number );
+
+			if ( ! $instance->already_exists() ) {
+				$instance->save();
+			}
+
+			unset( $instances_to_create[ $date_duration['original_index'] ] );
 			$this->processed++;
 		}
